requirements.txt
motor~=2.3.0
python-dotenv~=0.15.0
pymongo~=3.11.2
fastapi~=0.63.0
PyJWT~=2.0.1
pydantic~=1.7.3
uvicorn~=0.13.3